Convert to a feed view

+++ db/migrate/20151012021454_update_posts_default_call.rb
@@ -1,6 +1,6 @@
 class UpdatePostsDefaultCall < ActiveRecord::Migration
   def change
-    change_column :posts, :cost, :float, :default=>0
+    change_column :posts, :cost, :integer, :default=>0
     change_column_default(:posts, :call_to_action, nil)
   end
 end
